Add a proper build file
Conflicts:
	src/main/java/Movie.java
	src/main/java/Rental.java
	src/test/java/VideoStoreTest.java
