--- src/Movie.java
@@ -1,18 +1,0 @@
-
-
-public abstract class Movie
-{
-	private String title;
-
-	public Movie(String title) {
-		this.title 		= title;
-	}
-		
-	public String getTitle () {
-		return title;
-	}
-
-  public abstract double determineAmount(int daysRented);
-
-  public abstract int determineFrequentRenterPoints(int daysRented);
-}
