+++ src/main/java/Rental.java
@@ -1,0 +1,18 @@
+public class Rental {
+
+	public Rental (Movie movie, int daysRented) {
+		this.movie 		= movie;
+		this.daysRented = daysRented;
+	}
+	
+	public int getDaysRented () {
+		return daysRented;
+	}
+	
+	public Movie getMovie () {
+		return movie;
+	}
+	
+	private Movie movie;
+	private int daysRented;
+}
